# Quiz-Genie/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quivio - AI-Assisted Quiz Generation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List quizzes for the current user
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [faculty, student]
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a new quiz (faculty)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizBody"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get a quiz by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizWithQuestions"
        "404":
          description: Not found
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuizBody"
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/generate:
    post:
      operationId: generateQuizQuestions
      tags: [quizzes]
      summary: AI-generate questions for a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionsBody"
      responses:
        "200":
          description: Generated questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /quizzes/{id}/publish:
    post:
      operationId: publishQuiz
      tags: [quizzes]
      summary: Publish a quiz so students can attempt it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated quiz status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
...
components:
  schemas:
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount:
          type: integer
        duration:
          type: integer
          description: Duration in minutes
        status:
          type: string
          enum: [draft, published]
        createdBy:
          type: string
          description: Clerk user ID of creator
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        description:
          type: ["string", "null"]
        questionType:
          type: string
          enum: [multiple_choice, short_answer, mixed]
      required:
        - id
        - title
        - topic
        - difficulty
        - questionCount
        - duration
        - status
        - createdBy
        - createdAt
        - updatedAt
        - questionType
    QuizWithQuestions:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount:
          type: integer
        duration:
          type: integer
        status:
          type: string
          enum: [draft, published]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        description:
          type: ["string", "null"]
        questionType:
          type: string
          enum: [multiple_choice, short_answer, mixed]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required:
        - id
        - title
        - topic
        - difficulty
        - questionCount
        - duration
        - status
        - createdBy
        - createdAt
        - updatedAt
        - questionType
        - questions
    CreateQuizBody:
      type: object
      properties:
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount:
          type: integer
        duration:
          type: integer
        description:
          type: ["string", "null"]
        questionType:
          type: string
          enum: [multiple_choice, short_answer, mixed]
      required:
        - title
        - topic
        - difficulty
        - questionCount
        - duration
        - questionType
    UpdateQuizBody:
      type: object
      properties:
        title:
          type: string
        topic:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount:
          type: integer
        duration:
          type: integer
        description:
          type: ["string", "null"]
        questionType:
          type: string
          enum: [multiple_choice, short_answer, mixed]
        status:
          type: string
          enum: [draft, published]
    GenerateQuestionsBody:
      type: object
      properties:
        count:
          type: integer
        questionType:
          type: string
          enum: [multiple_choice, short_answer, mixed]
        additionalContext:
          type: ["string", "null"]
      required:
        - count
        - questionType
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, short_answer]
        options:
          type: ["string", "null"]
          description: JSON array of options for multiple choice
        correctAnswer:
          type: string
        explanation:
          type: string
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - quizId
        - questionText
        - questionType
        - correctAnswer
        - explanation
        - orderIndex
        - createdAt
...